+++ app/src/main/java/org/simple/clinic/home/HomeScreenEvent.kt
@@ -1,6 +1,12 @@
 package org.simple.clinic.home
 
 import org.simple.clinic.facility.Facility
+import org.simple.clinic.patient.Patient
+import org.simple.clinic.patient.businessid.Identifier
+import org.simple.clinic.scanid.EnteredShortCode
+import org.simple.clinic.scanid.ScanResult
+import org.simple.clinic.scanid.ScannedId
+import org.simple.clinic.util.Optional
 import org.simple.clinic.widgets.UiEvent
 
 sealed class HomeScreenEvent : UiEvent
@@ -12,3 +18,22 @@ object HomeFacilitySelectionClicked : HomeScreenEvent() {
 data class CurrentFacilityLoaded(val facility: Facility) : HomeScreenEvent()
 
 data class OverdueAppointmentCountLoaded(val overdueAppointmentCount: Int) : HomeScreenEvent()
+
+data class PatientSearchByIdentifierCompleted(val patient: Optional<Patient>, val identifier: Identifier) : HomeScreenEvent()
+
+sealed class BusinessIdScanned : HomeScreenEvent() {
+
+  companion object {
+    fun fromScanResult(scanResult: ScanResult): BusinessIdScanned {
+      return when (scanResult) {
+        is ScannedId -> ByIdentifier(scanResult.identifier)
+        is EnteredShortCode -> ByShortCode(scanResult.shortCode)
+      }
+    }
+  }
+
+  data class ByIdentifier(val identifier: Identifier) : BusinessIdScanned()
+
+  data class ByShortCode(val shortCode: String) : BusinessIdScanned()
+}
+