+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Internal
 - [In Progress: 11 Nov 2020] Migrate app screens to use ViewBinding
 - Bump appcompat -> 1.2.0
+- Move BP passport scan results handling to `HomeScreen`
 
 ### Changes
 - Disable the change language feature on devices running Lollipop (API level 21, 22)